trained a smaller model as proof of concept.
